fix(search pill): define behavior for potential name edge case
If a search pill name is created by selecting an option from a potential
name and then the potential names are removed (e.g. by changing a filter
in the tasklist), the name field is treated as text input, but still
contains an object. This commit handles this edge case

related to CAM-3902